Add files via upload
to make it work you should put the database in the same directory as the pythonfile. the main password to login is 'test' if you want to add a new password click add, normally everything is encrypted and safe, if you find any flaws in the code please tell me so i can improve.